Improved types (#4)
* Improved typings for model data represented as unions 
* Increased readability of merged types
* Added downcast option to scan/query/get/update

If a model represents multiple different types of data, as is common with NOSQL modelling, then the type T of a model will typically be a union type.
Due to how the standard Pick and Omit types do not distribute over unions, this caused some problems, since all calls returning data used `Pick<T, P>` which does not distribute, resulting in returning only the attributes found in all possible types of the union. This is now fixed using a distributed type, and by using `null` to denote "full projection", all the ugly return types such as `Pick<Foo, 'a'|'b'|'c'|'d'|'e'|'f'>` for the default full projections are avoided and the return type for a full projection is simply `Foo`.

Also, added a way to explicitly downcast data when knowing that this is safe.
For example, if a table contains two types of data - `Foo` and `Bar` distinguishable by an `itemType` discriminator attribute,  and the model is then defined as `Model<Foo | Bar>`, and querying for `itemType: 'foo'`, then it's now possible to request that the returned items are of type `Foo` instead of `Foo | Bar`:

```
const { items } = await model.query({
  type: as<Foo>(),
  keyCondition: {
    itemType: 'foo'
  }
};

// typeof items is Foo[]
```

This is essentially the same as 

```
const res = await model.query({
  keyCondition: {
    itemType: 'foo'
  }
};

const items = res.items as Foo[];
```

This avoids having to downcast data separately, which is cumbersome with destructuring, and makes it clearer since the type hint is passed in the same parameters as the key condition that effectively implements the filtering of items to the specific type given.

Furthermore, If using `withCreator()` to automatically add attributes such as timestamps, the types are now clearer, especially if the model data type already includes the timestamp attributes:

```
interface TimestampAttributes {
  createdAt: string;
  updatedAt: string;
}

interface Foo extends TimestampAttributes {
  id: string;
  ...
}

class FooModel extends DynamoClient.model<Foo>()
  .withKey('id')
  .withCreator<TimestampAttributes>(item => ({createdAt: now(), updatedAt: now()}))
  .withUpdater(item => Object.assign(item, {updatedAt: now()}))
  .class() {}
```

Now, since `Foo` already extends `TimestampAttributes`, the type `T` of the model will simply be `Foo` instead of `Foo & TimestampAttributes`, which makes working with the model clearer.

and properly distributes across unions.